Enable ReadMemory in the code window
In the code window, ask the user for the memory address.  This allows
a broader "Memory View".

If the clipboard @" seems to contain a valid address, use that as the
default in the prompt.  This way, if the user has yanked the value of
a pointer, we can easily inspect its memory; the variable Dump would
show bytes around the variable holding the pointer.

Change VimspectorBalloonEval to yank the highlighted line when the
user accepts the pop up selection.  Default key is <Space>, add
'c' and 'C'.

All together, this means the user can put the cursor on a pointer,
use the balloon to see its value, press 'C', and then call
ReadMemory to see the memory it points at.

Note: not tested in neovim, not tested with the mouse, no coverage
in unittests.